requirements.txt
# Main requirements for LearnPodcast project
flask>=2.0.0
flask-login>=0.6.0
flask-sqlalchemy>=3.0.0
gunicorn>=20.0.0
numpy>=2.2.5
pandas>=2.2.3
psycopg2-binary>=2.9.10
scikit-learn>=1.6.1
statsmodels>=0.14.4
tensorflow>=2.14.0
textblob>=0.19.0
trafilatura>=2.0.0
tweepy>=4.15.0
yfinance>=0.2.58
requests>=2.32.3
lxml>=5.4.0
python-dateutil>=2.9.0
protobuf>=4.25.7
markupsafe>=3.0.2
soupsieve>=2.7
python-dotenv>=1.0.0
beautifulsoup4
curl-cffi
frozendict
multitasking
peewee
platformdirs
pytz
oauthlib
requests-oauthlib
feedparser>=6.0.10
html5lib>=1.1
nltk>=3.8.1
absl-py
astunparse
flatbuffers
gast
google-pasta
grpcio>=1.71.0
h5py
keras
libclang
ml-dtypes
opt-einsum
packaging
setuptools
six
tensorboard
tensorflow-estimator
tensorflow-io-gcs-filesystem
termcolor
typing-extensions
wrapt
joblib
threadpoolctl
pyasn1
rsa
nltk
python-dotenv>=1.0.0
